+++ src/entity/semVer.ts
@@ -1,0 +1,3 @@
+export function validRange(str) {
+    return /^[<>~=]/.test(str) && str.includes('*')
+}