+++ src/extension.ts
@@ -1,5 +1,6 @@
 import { DocumentSelector, ExtensionContext, languages, ShellExecution, Task, TaskDefinition, tasks, TaskScope, window, workspace } from 'vscode'
 import { CratesCompletionProvider, Listener, rustAnalyzer } from '@/controller'
+import { cargoTomlAction } from './controller/codeAction'
 
 // This method is called when your extension is activated
 // Your extension is activated the very first time the command is executed
@@ -42,6 +43,7 @@ export async function activate(context: ExtensionContext) {
 
 		workspace.createFileSystemWatcher("**/Cargo.lock").onDidChange(listener.onDidLockFileChange, listener),
 
+		languages.registerCodeActionsProvider(documentSelector, cargoTomlAction),
 
 		// Register our versions completions provider
 		languages.registerCompletionItemProvider(